13c NPC Loot Drop / PC Pickup
Allow the player to pick up and "equip" whatever the NPC drops. The issue is that it breaks our guns when we switch to one of the ones we pick up or directly on pickup. I think this will be fixed in a future video and isn't an issue with something I did.