Complete Replacement
One of the first movement bugs I ran into in the course was improperly fixed. I remember emailing Jason about it and my fix was to reorder something in the code. Well it finally is starting to cause errors with placement in the hotbar (I think) and also item pickups. Since I don't remember what I did nor do I have the time to go back to the first couple sections of the course to find out I just have completely replaced the project with the snapshot from 20d. So far this seems to be working out fine. My original project has been saved so I can copy comments and debug at a later date.